+++ src/core/colors.cpp
@@ -100,9 +100,10 @@ namespace PixelMaestro {
 			case MixMode::Alpha:
 			{
 				float alpha_pct = alpha / (float)255;
-				mixed_color.r = (alpha_pct * color_two.r) + ((1 - alpha_pct) * color_one.r);
-				mixed_color.g = (alpha_pct * color_two.g) + ((1 - alpha_pct) * color_one.g);
-				mixed_color.b = (alpha_pct * color_two.b) + ((1 - alpha_pct) * color_one.b);
+				float alpha_pct_inv = 1 - alpha_pct;
+				mixed_color.r = (alpha_pct * color_two.r) + (alpha_pct_inv * color_one.r);
+				mixed_color.g = (alpha_pct * color_two.g) + (alpha_pct_inv * color_one.g);
+				mixed_color.b = (alpha_pct * color_two.b) + (alpha_pct_inv * color_one.b);
 				break;
 			}
 			case MixMode::Multiply: