+++ src/core/section.h
@@ -184,7 +184,7 @@ namespace PixelMaestro {
 			Point* get_offset();
 			Section* get_parent_section() const;
 			Pixel* get_pixel(uint16_t x, uint16_t y) const;
-			Colors::RGB get_pixel_color(uint16_t x, uint16_t y);
+			Colors::RGB get_pixel_color(uint16_t x, uint16_t y, Colors::RGB* base_color = nullptr);
 			Scroll* get_scroll() const;
 			void remove_animation();
 			void remove_canvas();