+++ src/animation/sparkleanimation.h
@@ -6,7 +6,7 @@
 namespace PixelMaestro {
 	class SparkleAnimation : public Animation {
 		public:
-			SparkleAnimation(Section* section);
+			explicit SparkleAnimation(Section* section);
 			~SparkleAnimation();
 			uint8_t get_threshold() const;
 			void set_threshold(uint8_t threshold);