+++ CHANGELOG.md
@@ -21,7 +21,7 @@ WARNING: This update breaks backwards compatability with v0.12 and earlier Cuefi
 - Animations
 	- Removed vertical orientation for Mandelbrot animations.
 - Layers
-	- Rewrote Layer mixing algorithm. This fixes an issue when combining the Overlay and Alpha MixModes in Sections with 2+ Layers.
+	- Rewrote Layer mixing algorithm. This fixed an issue when combining the Overlay and Alpha MixModes in Sections with more than one Layer.
 	
 ### Removed
 - Animations